[modify] modify packages and add some annotations.
